+++ fieldmeter.h
@@ -43,8 +43,8 @@ protected:
   enum UsedType { INVALID_0, FLOAT, PERCENT, AUTOSCALE, INVALID_TAIL };
 
   int numfields_;
-  float *fields_;
-  float total_, used_, lastused_;
+  double *fields_;
+  double total_, used_, lastused_;
   int *lastvals_, *lastx_;
   unsigned long *colors_;
   unsigned long usedcolor_;
